configurable settings in DB

+++ lib/mix/tasks/pleroma/config.ex
@@ -19,38 +19,50 @@ defmodule Mix.Tasks.Pleroma.Config do
   def run(["migrate_to_db"]) do
     Common.start_pleroma()
 
-    Application.get_all_env(:pleroma)
-    |> Enum.reject(fn {k, _v} -> k in [Pleroma.Repo, :env] end)
-    |> Enum.each(fn {k, v} ->
-      key = to_string(k) |> String.replace("Elixir.", "")
-      {:ok, _} = Config.update_or_create(%{key: key, value: v})
-      Mix.shell().info("#{key} is migrated.")
-    end)
+    if Pleroma.Config.get([:instance, :dynamic_configuration]) do
+      Application.get_all_env(:pleroma)
+      |> Enum.reject(fn {k, _v} -> k in [Pleroma.Repo, :env] end)
+      |> Enum.each(fn {k, v} ->
+        key = to_string(k) |> String.replace("Elixir.", "")
+        {:ok, _} = Config.update_or_create(%{key: key, value: v})
+        Mix.shell().info("#{key} is migrated.")
+      end)
 
-    Mix.shell().info("Settings migrated.")
+      Mix.shell().info("Settings migrated.")
+    else
+      Mix.shell().info(
+        "Migration is not allowed by config. You can change this behavior in instance settings."
+      )
+    end
   end
 
   def run(["migrate_from_db", env]) do
     Common.start_pleroma()
 
-    config_path = "config/#{env}.migrated.secret.exs"
+    if Pleroma.Config.get([:instance, :dynamic_configuration]) do
+      config_path = "config/#{env}.migrated.secret.exs"
 
-    {:ok, file} = File.open(config_path, [:write])
+      {:ok, file} = File.open(config_path, [:write])
 
-    Repo.all(Config)
-    |> Enum.each(fn config ->
-      mark = if String.starts_with?(config.key, "Pleroma."), do: ",", else: ":"
+      Repo.all(Config)
+      |> Enum.each(fn config ->
+        mark = if String.starts_with?(config.key, "Pleroma."), do: ",", else: ":"
 
-      IO.write(
-        file,
-        "config :pleroma, #{config.key}#{mark} #{inspect(Config.from_binary(config.value))}\r\n"
+        IO.write(
+          file,
+          "config :pleroma, #{config.key}#{mark} #{inspect(Config.from_binary(config.value))}\r\n"
+        )
+
+        {:ok, _} = Repo.delete(config)
+        Mix.shell().info("#{config.key} deleted from DB.")
+      end)
+
+      File.close(file)
+      System.cmd("mix", ["format", config_path])
+    else
+      Mix.shell().info(
+        "Migration is not allowed by config. You can change this behavior in instance settings."
       )
-
-      {:ok, _} = Repo.delete(config)
-      Mix.shell().info("#{config.key} deleted from DB.")
-    end)
-
-    File.close(file)
-    System.cmd("mix", ["format", config_path])
+    end
   end
 end
